LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/api/basic/scheme-about.any-expected.txt

CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:blank due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:invalid.com due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:config due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:unicorn due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: Fetch API cannot load about:blank due to access control checks.

PASS Fetching about:blank with method GET is KO
PASS Fetching about:blank with method PUT is KO
PASS Fetching about:blank with method POST is KO
PASS Fetching about:invalid.com with method GET is KO
PASS Fetching about:config with method GET is KO
PASS Fetching about:unicorn with method GET is KO
PASS Fetching about:blank with range header does not affect behavior
